fastapi_app/requirements.txt
# FastAPI SOAR/SIEM Platform - Requirements
# Python 3.10+

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database - PostgreSQL
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.14.0

# Database - ClickHouse
clickhouse-connect>=0.7.0

# Configuration
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Template Engine
jinja2>=3.1.2

# Utilities
python-multipart>=0.0.6
lz4>=4.3.0
zstandard>=0.22.0
pytz>=2023.3
openpyxl>=3.1.0

# HTTP Client (for testing)
httpx>=0.25.0

# Production Server
gunicorn>=21.2.0

# SSH for device management
paramiko>=3.4.0
asyncssh>=2.14.0

# Encryption for credentials
cryptography>=41.0.0

# Background tasks / scheduling
apscheduler>=3.10.0

# Authentication
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# Appliance CLI
prompt_toolkit>=3.0.0
pyyaml>=6.0
